Perform early emerge --depclean and genup update check
We need do do an "emerge --depclean" immediately after the
@world update, to remove obsolete pacakges that may otherwise
end up being rebuilt (unnecessarily) by the @preserved-rebuild
emerge stage.

Also, since it is quick to check (and may be necessary to
fix important bugs, that would otherwise prevent the system
from updating) we now attempt to update genup itself early in
the process (and, if this causes the reported version number
to change, exec the new copy).